Task4

+++ src/main/java/com/es/phoneshop/model/cart/Cart.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 public class Cart implements Serializable {
     private List<CartItem> cartItems = new ArrayList<>();
     private BigDecimal totalCost = new BigDecimal(0);
-    private Long totalQuantity = 0L;
+    private int totalQuantity = 0;
 
     @Override
     public boolean equals(Object o) {
@@ -41,11 +41,11 @@ public class Cart implements Serializable {
         this.totalCost = totalCost;
     }
 
-    public Long getTotalQuantity() {
+    public int getTotalQuantity() {
         return totalQuantity;
     }
 
-    public void setTotalQuantity(Long totalQuantity) {
+    public void setTotalQuantity(int totalQuantity) {
         this.totalQuantity = totalQuantity;
     }
